+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Support Reduct Storage API v0.6, [PR-29](https://github.com/reduct-storage/reduct-js/pull/29)
 * Add data streaming, [PR-30](https://github.com/reduct-storage/reduct-js/pull/30)
+* Add original exception to APIError, [PR-31](https://github.com/reduct-storage/reduct-js/pull/31)
 
 **Docs**: